Restructure the test file
